+++ editors/shared/tree-sitter-toon/queries/highlights.scm
@@ -2,6 +2,7 @@
 
 ; Comments
 (comment) @comment
+(block_comment) @comment
 
 ; Keys (property names)
 (key) @property
@@ -9,6 +10,7 @@
 ; Strings
 (double_quoted_string) @string
 (single_quoted_string) @string
+(block_string) @string
 (unquoted_string) @string
 
 ; Escape sequences are inline in string tokens